TransformExtensions.AbsoluteMaxValue
