+++ Assets/UniGLTF/Runtime/SpringBoneJobs/SpringBoneJobs.asmdef
@@ -1,7 +1,9 @@
 {
     "name": "SpringBoneJobs",
     "rootNamespace": "",
-    "references": [],
+    "references": [
+        "GUID:8d76e605759c3f64a957d63ef96ada7c"
+    ],
     "includePlatforms": [],
     "excludePlatforms": [],
     "allowUnsafeCode": false,
