+++ Assets/VRM/Runtime/SpringBone/Logic/SphereCollider.cs
@@ -1,3 +1,4 @@
+using UniGLTF;
 using UnityEngine;
 
 namespace VRM.SpringBone
