+++ Assets/VRM/Runtime/SpringBone/Logic/SpringBoneJointInit.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using UniGLTF;
 using UnityEngine;
 
 namespace VRM.SpringBone
