+++ Assets/VRM/Runtime/SpringBone/Logic/SpringBoneSystem.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
+using UniGLTF;
 using UnityEngine;
 
 namespace VRM.SpringBone
